requirements.txt
# Core Data Science Libraries
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Advanced ML Models
lightgbm==4.1.0
xgboost==2.0.3
catboost==1.2.2

# Model Selection & Optimization
optuna==3.5.0
scikit-optimize==0.9.0

# API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Utilities
python-dotenv==1.0.0
joblib==1.3.2
tqdm==4.66.1
pyyaml==6.0.1

# Data Processing
imbalanced-learn==0.11.0
shap==0.44.0
